pyproject.toml
[project]
name = "gpu-docker-test"
version = "0.1.0"
description = "GPU test in Docker with UV-managed dependencies"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x",
    "numpy<2.0",
    "psutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["src"]